Implement Flexible Logic for Saving Assessments in the Database
The API should provide flexibility for clients to either use an existing Subject by specifying its id in the db_id property or create a new one by leaving db_id empty and filling in the other three properties (name, type and id).

Finally, the API should give precedence to the properties stored in the database. If a valid db_id is provided, its associated Subject should be retrieved, and the values of name, type, and id in the Assessment should be overwritten with the corresponding values from the existing  database Subject.